logging and fixing postMessage
As I've been learning more about hooks and redux I've adding some additional console logging to better understand the flow of data.

The postMessage function inside thunkCreators.js was async as was saveMessage, the function it calls, but the return value wasn't await so I added that.

After doing that I fixed how the reducer function receives that data. Now the correct data goes to addMessageToStore.