Fixing the order of messages
Now the Messages in each conversation have been re-ordered so that the oldest is on top, and the latest is on bottom (closer to the input bar).

I did this by changing the messages array within the greater data object, then I sent that new version of data to the dispatch just like the previous one.