making reducer functions immutable
I modified addMessageToStore and addNewConvoToStore so that the return a copy of the conversations to make an immutable change to state rather than mutating it directly.

This allows me to remove the useSelector in ActiveChat, and for the ActiveChat component to still update when changes to the store are made.