Update db.js
The Tests are still failing on Cypress. I see errors about the connection being refused at 127.0.0.1:3000. I'm thinking that might not be a smart way for me to connect to the database since it is possible to be running on a different local IP. Using Localhost seems like a more universal option.