refactor of signup
I moved most of the component rendering away from signup.js so that I could conditionally render based on screen size in a more clean way.
I can now re-style and have more accurate useStyles for each component, even if it's the same component, but for a different screen size.