Fixing structure
After doing some research on best practices I came up with a different method to dynamically change the layout of the grid, and make sure that I can still have a landing image on all versions that looks good.
There are still a few things that aren't quite right with the styling but it's a lot better than it was last commit.